MR comments - api/ prefix comes from base-service - place-types service - zoom out a little when closing place details - map stays mounted when switching to view details to reduce loading time when returning to map
