+++ SportMap.AL/DTOs/PlaceDto.cs
@@ -40,13 +40,4 @@ namespace SportMap.AL.DTOs
         public DateTime? UpdatedAt { get; init; }
         public string Status { get; init; } = string.Empty;
     }
-
-    public class PlaceTypeDto : IDTO
-    {
-        public PlaceTypeDto() { }
-
-        public Guid Id { get; set; }
-        public string Name { get; init; } = string.Empty;
-        public string Description { get; init; } = string.Empty;
-    }
 }
