+++ frontend/src/components/MapView.tsx
@@ -91,6 +91,8 @@ export default function MapView({ places, selectedPlace: selectedPlaceProp, onPl
   };
 
   const handlePlaceClose = () => {
+    const map = mapInstanceRef.current;    
+    map.zoomOut();    
     setSelectedPlace(null);
     if (onPlaceSelect) {
       onPlaceSelect(null);
